datasets/ImageDatasets/requirements.txt
torch
torchvision
boto3
pymongo
zmq
numpy
pillow